feat(treasury): add swap functionalities through osmosis  (#126)
## Overview

This PR updates the treasury and staking contracts, allowing the staking
contract to send protocol fees to the treasury and enabling a `trader`
account to perform swap operations, helping the treasury hedge against
asset price fluctuations.

To prevent the `trader` from converting the treasury's assets into
unwanted ones, a list of swap routes can be defined. This forces the
`trader` to use only a predefined set of pools and only in a predefined
direction. For example, if the Osmosis pool 1 has the OSMO/USDC pair, we
can enforce that the `trader` can only swap OSMO for USDC, but not the
other way around.

closes: MILKTIA-35

## What changes have been made in this PR?

- [ ] Add swap functionalities to the treasury contract
- [ ] Allow the staking contract to send the protocol fees to the
treasury
- [ ] Fix lint errors reported by `cargo clippy` 

## Checklist

---

- [ ] Appropriate labels applied
- [ ] Targeted PR against correct branch
- [ ] Linked to Github issue with discussion and accepted design OR link
to spec that describes this work.
- [ ] Wrote unit and integration
- [ ] Updated relevant documentation