feat(staking): add miniwasm x/tokenfactory support (#135)
<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰
v    Before smashing the submit button please review the checkboxes.
v If a checkbox is n/a - please still include it but + a little note why
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## Overview

Closes: LSTS-87

This PR updates the staking contract to be compatible with the
`x/tokenfactory` module in
[miniwasm](https://github.com/initia-labs/miniwasm).
Compatibility is achieved through the `miniwasm` feature, which, when
enabled, includes the `x/tokenfactory` proto definitions from miniwasm.

## Checklist

---

- [ ] Appropriate labels applied
- [ ] Targeted PR against correct branch
- [ ] Linked to Github issue with discussion and accepted design OR link
to spec that describes this work.
- [ ] Wrote unit and integration
- [ ] Updated relevant documentation