fix: update oracle price after contract resume (#136)
<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰
v    Before smashing the submit button please review the checkboxes.
v If a checkbox is n/a - please still include it but + a little note why
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## Overview

This PR fixes a bug where we send the wrong redeemption rate to the
oracle when the contract is resumed.
This is caused by the fact that `update_oracle_msgs` internally uses the
contract `STATE` to compute the redemption rate but during the contract
resume logic the state is written after the invocation of
`update_oracle_msgs` causing the contract to send the wrong redemptio
rate to the oracle contract.
 
closes: #XXXX

## What changes have been made in this PR?

- [ ]

## Checklist

---

- [ ] Appropriate labels applied
- [ ] Targeted PR against correct branch
- [ ] Linked to Github issue with discussion and accepted design OR link
to spec that describes this work.
- [ ] Wrote unit and integration
- [ ] Updated relevant documentation