feat(staking): improve contract initializtion (#144)
<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰
v    Before smashing the submit button please review the checkboxes.
v If a checkbox is n/a - please still include it but + a little note why
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## Overview

This PR aims to reduce the number of transactions required to deploy the
`staking` contract. Since the `oracle` contract requires the
`admin_address` to be the `staking` contract, we need to know the
`staking` contract address in order to initialize the `oracle` contract.

This PR allows specifying the code ID of the `oracle` contract so that
both contracts can be initialized with a single `instantiate`
transaction.

Closes: LSTS-118

## What changes have been made in this PR?

- [ ]

## Checklist

---

- [ ] Appropriate labels applied
- [ ] Targeted PR against correct branch
- [ ] Linked to Github issue with discussion and accepted design OR link
to spec that describes this work.
- [ ] Wrote unit and integration
- [ ] Updated relevant documentation